+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+- Fix date format prasing issue, closes Github issue [#333](https://github.com/EasyPost/easypost-java/issues/333)
+
 ## v7.4.2 (2024-08-09)
 
 - Fix pagination parameters for `getNextPage` in `User` service